requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-jose==3.3.0
pydantic==2.7.0
pytest==7.4.0
httpx==0.24.0
websockets==11.0.3
numpy
scipy
praat-parselmouth
spacy
pyannote.audio
torch